--- src/main/java/org/owasp/dsomm/metricca/analyzer/controller/GrafanaTeamController.java
@@ -1,27 +1,0 @@
-package org.owasp.dsomm.metricca.analyzer.controller;
-
-import org.eclipse.jgit.api.errors.GitAPIException;
-import org.owasp.dsomm.metricca.analyzer.deserialization.Application;
-import org.owasp.dsomm.metricca.analyzer.deserialization.ApplicationDirector;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.io.IOException;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
-
-@Controller
-public class GrafanaTeamController {
-  private static final Logger logger = LoggerFactory.getLogger(GrafanaTeamController.class);
-
-  @Autowired
-  private ApplicationDirector applicationDirector;
-
-
-}
