+++ src/main/java/org/owasp/dsomm/metricca/analyzer/controller/GrafanaTeamDashboardController.java
@@ -1,11 +1,21 @@
 package org.owasp.dsomm.metricca.analyzer.controller;
 
+import org.owasp.dsomm.metricca.analyzer.deserialization.Application;
 import org.owasp.dsomm.metricca.analyzer.deserialization.ApplicationDirector;
+import org.owasp.dsomm.metricca.analyzer.deserialization.activity.Activity;
+import org.owasp.dsomm.metricca.analyzer.grafana.PanelConfiguration;
+import org.owasp.dsomm.metricca.analyzer.grafana.PanelFactory;
 import org.owasp.dsomm.metricca.analyzer.grafana.TeamDashboard;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @Controller
 public class GrafanaTeamDashboardController {
@@ -13,9 +23,6 @@ public class GrafanaTeamDashboardController {
 
   @Autowired
   private ApplicationDirector applicationDirector;
-
-  @Autowired
-  private TeamDashboard teamDashboard;
 
 
 }
